+++ CHANGELOG.md
@@ -17,6 +17,7 @@ This release has an [MSRV][] of 1.86.
 
 ### Added
 
+- `AlphaColor::<Srgb>::from_hex` and `OpaqueColor::<Srgb>::from_hex`, which can be used to convert const colors from an in-editor color picker. ([#217][] by [@l0uisgrange][] and [@DJMcNab][])
 - Add a `const` `DynamicColor::new` constructor for convenience, taking a color space tag and color components, and setting default `Flags`. ([#219][] by [@tomcur][])
 
 ### Changed
@@ -175,6 +176,7 @@ This is the initial release.
 [@ajakubowicz-canva]: https://github.com/ajakubowicz-canva
 [@alvinisspicy]: https://github.com/alvinisspicy
 [@DJMcNab]: https://github.com/DJMcNab
+[@l0uisgrange]: https://github.com/l0uisgrange
 [@LaurenzV]: https://github.com/LaurenzV
 [@MightyBurger]: https://github.com/MightyBurger
 [@raphlinus]: https://github.com/raphlinus
@@ -234,6 +236,7 @@ This is the initial release.
 [#202]: https://github.com/linebender/color/pull/202
 [#210]: https://github.com/linebender/color/pull/210
 [#211]: https://github.com/linebender/color/pull/211
+[#217]: https://github.com/linebender/color/pull/217
 [#218]: https://github.com/linebender/color/pull/218
 [#219]: https://github.com/linebender/color/pull/219